requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy>=2.0
redis>=5.0
yt-dlp
faster-whisper
sentence-transformers
numpy

boto3
google-cloud-storage
google-auth
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Pillow
openai>=1.30.0
tweepy
prometheus_client>=0.20.0